refactor: 계층형으로 auth 패키지 변경
