+++ src/main/java/cart/resolver/Authentication.java
@@ -1,4 +1,4 @@
-package cart.auth.resolver;
+package cart.resolver;
 
 import static java.lang.annotation.ElementType.PARAMETER;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
